+++ ios/PresentationLayer/Sources/PresentationLayer/App/Tab1-Users/Detail/UserDetailView.swift
@@ -1,0 +1,18 @@
+//
+//  Created by Patrik Cesnek on 04.03.2022
+//  Copyright © 2022 Matee. All rights reserved.
+//
+
+import SwiftUI
+
+struct UserDetailView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct SwiftUIView_Previews: PreviewProvider {
+    static var previews: some View {
+        UserDetailView()
+    }
+}
